+++ apps/room/src/lib/components/PresentationArea.svelte
@@ -711,6 +711,7 @@
                       {doNotDisturbOn}
                       {bufferSizeLevel}
                       {onBufferSizeChange}
+                      {menus}
                     />
                   </div>
                 </div>